HS-106 FEAT: 유효성 검사 함수 완성
