+++ Makefile
@@ -27,7 +27,7 @@ UTILS_SRCS = env_utils/env_utils.c env_utils/env_key_valid_checker.c env_utils/g
 	  env_utils/make_node_to_envp.c  \
 	  built_in_utils/cd_utils.c \
 	  replace_dollar/replace_dollar.c replace_dollar/replace_dollar_append_utils.c replace_dollar/replace_dollar_len_utils.c \
-	  chore_utils/is_same_string.c chore_utils/safe_malloc.c  \
+	  chore_utils/is_same_string.c chore_utils/safe_malloc.c chore_utils/quote_validator.c \
 	  free_utils/free_utils.c \
 	  signal.c get_token_head.c ft_error.c word_check.c
 UTILS_PATH = $(addprefix $(UTILS_DIR), $(UTILS_SRCS))
